requirements.txt
random==1.0.1
tensorflow==2.6.0
pandas==1.3.3
typing==3.7.4.3
argparse==1.4.0
python-dotenv==0.19.1
keras==2.6.0